+++ assets/javascripts/kitten/components/text-copy/test.js
@@ -1,0 +1,25 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { TextCopy } from './index'
+
+describe('<TextCopy />', () => {
+  it('should match snapshot', () => {
+    const component = renderer
+      .create(<TextCopy alertMessage="Text copied!">Text To Copy</TextCopy>)
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
+  it('should match snapshot with description', () => {
+    const component = renderer
+      .create(
+        <TextCopy
+          alertMessage="Text copied!"
+          description="This is a description"
+        >
+          Text To Copy
+        </TextCopy>,
+      )
+      .toJSON()
+    expect(component).toMatchSnapshot()
+  })
+})